+++ tests/generic/models.py
@@ -47,6 +47,8 @@ else:
     GenericForeignKey = None
 
 GENDER_CHOICES = [
+    (None, ""),
+    ("M", "male"),
     ("M", "male"),
     ("F", "female"),
     ("N", "non-binary"),